decode: bounds-check the untrusted stream
The decoder previously ignored in_len entirely and trusted K_right from
the wire: a truncated or malicious stream could walk in_ptr past the
input buffer, and a K_right > K made K_left negative, feeding negative
sizes to the scratch carver and merges (out-of-bounds writes).

Every stream read is now validated before dereferencing — the block-N
header against in_len, K_right headers and bitmap records against
in_end (wire_read_kr_checked / wire_read_bitmap_checked), flat packed-
bit regions against in_end — and K_right is required to be in [0, K],
which keeps every child size, carve and merge extent inside the bounds
the entry sized for N.  The FSE record path now also honors
pivco_fse_decompress's error return (bad table id / malformed payload
-> CORRUPT) instead of consuming an uninitialized bitmap.  A bitmap
whose popcount disagrees with K_right still yields garbage output, but
provably touches only in-bounds memory (merges are extent-bounded).
All checks are per NODE record — a couple of predictable compares each,
~100 per 16K block; decode bandwidth is unchanged on M4.

extras/fuzz_decode.c: 300K corrupted-stream iterations (flips,
truncations, garbage) with canary bands around the output — guards
intact, every return is OK or CORRUPT.  ASan+UBSan on the M4 mini:
clean over tests + 60K fuzz iterations, after fixing two pre-existing
UBSan signed-left-shift complaints in the NEON merge primitives
(mask << 4 with the top bitmap bit set; same bits, now unsigned).

+++ src/pivco_huffman_primitives_neon.h
@@ -121,8 +121,8 @@ static inline void merge_neon_16B(uint8_t *dest, const uint8_t *l_list,
                                   const uint8_t *r_list, intptr_t mask,
                                   const int8_t *tab0, const int8_t *tab1)
 {
-    int8x16_t shuf0 = vld1q_s8(&tab0[(mask << 4) & 0xff0]);
-    int8x16_t shuf1 = vld1q_s8(&tab1[(mask >> 4) & 0xff0]);
+    int8x16_t shuf0 = vld1q_s8(&tab0[((uintptr_t)mask << 4) & 0xff0]);
+    int8x16_t shuf1 = vld1q_s8(&tab1[((uintptr_t)mask >> 4) & 0xff0]);
     uint8x16_t shuf = vreinterpretq_u8_s8(vabdq_s8(shuf0, shuf1));
     uint8x16x2_t src;
     src.val[0] = vld1q_u8(r_list);
@@ -225,8 +225,8 @@ static inline void merge_cst_vec_neon(const uint8_t *bm, int K,
         uint64_t pfx = vget_lane_u64(vreinterpret_u64_u8(pop8), 0) * 0x0101010101010101ull;
         intptr_t p0 = (pfx >> 8) & 0xff, p1 = (pfx >> 24) & 0xff, p2 = (pfx >> 40) & 0xff, p3 = pfx >> 56;
 #define _MCV(off, rd, mk) do {                                                   \
-        int8x16_t s0 = vld1q_s8(&g_merge_shuf0[(((intptr_t)(mk)) << 4) & 0xff0]); \
-        int8x16_t s1 = vld1q_s8(&g_merge_shuf1[(((intptr_t)(mk)) >> 4) & 0xff0]); \
+        int8x16_t s0 = vld1q_s8(&g_merge_shuf0[(((uintptr_t)(mk)) << 4) & 0xff0]); \
+        int8x16_t s1 = vld1q_s8(&g_merge_shuf1[(((uintptr_t)(mk)) >> 4) & 0xff0]); \
         uint8x16_t sh = vreinterpretq_u8_s8(vabdq_s8(s0, s1));                    \
         uint8x16x2_t src; src.val[0] = vld1q_u8(rd); src.val[1] = Lb;            \
         vst1q_u8(out + i + (off), vqtbl2q_u8(src, sh));                          \
